+++ packages/@baldeweg/ui/src/components/BForm/BFormFieldset.spec.js
@@ -1,9 +1,0 @@
-import { describe, it, expect } from 'vitest'
-
-import BFormFieldset from '../BForm/BFormFieldset.vue'
-
-describe('BFormFieldset', () => {
-  it('shows BFormFieldset', () => {
-    expect(BFormFieldset).toBeTruthy()
-  })
-})