+++ packages/@baldeweg/ui/src/components/BForm/BFormGroup.spec.js
@@ -1,9 +1,0 @@
-import { describe, it, expect } from 'vitest'
-
-import BFormGroup from '../BForm/BFormGroup.vue'
-
-describe('BFormGroup', () => {
-  it('shows BFormGroup', () => {
-    expect(BFormGroup).toBeTruthy()
-  })
-})