+++ packages/@baldeweg/ui/src/components/BForm/BFormLabel.vue
@@ -1,5 +1,0 @@
-<template>
-  <label>
-    <slot />
-  </label>
-</template>